Configure Apollo local state
