+++ client/src/App.js
@@ -1,19 +1,8 @@
 import React, { Component } from "react";
 import { ApolloProvider } from "react-apollo";
-import { ApolloClient } from "apollo-client";
-import { InMemoryCache } from "apollo-cache-inmemory";
-import { HttpLink } from "apollo-link-http";
 
 import Main from "./components/Pages/Main";
-
-const cache = new InMemoryCache();
-const link = new HttpLink({
-	uri: "http://localhost:4000/graphql"
-});
-const client = new ApolloClient({
-	cache,
-	link
-});
+import client from "./client";
 
 class App extends Component {
 	render() {
